web-ui/backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
pydantic
pydantic-settings
psutil
py-cpuinfo
